+++ tests/isbase64.test.js
@@ -1,4 +1,35 @@
+import { randomBytes } from "node:crypto";
 import { isBase64 } from "../dist/ch";
+
+function randB64(length = 32) {
+	const b64Secret = randomBytes(length).toString("base64");
+	// Replace '+' with '-', '/' with '_', and remove '=' padding
+	return b64Secret.replace(/\+/g, "-").replace(/\//g, "_").replace(/=+$/, "");
+}
+const sec1 = randB64(30);
+const sec2 = randB64(32);
+const sec3 = randB64(34);
+const sec4 = randB64(36);
+console.log("sec1", sec1);
+console.log("sec2", sec2);
+console.log("sec3", sec3);
+console.log("sec4", sec4);
+
+test("sends random base64 string sec1 to isBase64 urlEncoded = true", () => {
+  expect(isBase64(sec1, true)).toBe(true);
+});
+
+test("sends random base64 string sec2 to isBase64 urlEncoded = true", () => {
+  expect(isBase64(sec2, true)).toBe(true);
+});
+
+test("sends random base64 string sec3 to isBase64 urlEncoded = true", () => {
+  expect(isBase64(sec3, true)).toBe(true);
+});
+
+test("sends random base64 string sec4 to isBase64 urlEncoded = true", () => {
+  expect(isBase64(sec4, true)).toBe(true);
+});
 
 test("sends null to isBase64 urlEncoded = false", () => {
   expect(isBase64(null)).toBe(false);